+++ test/cases/_preprocessor/js-tmpl-hbs-compile-helpers-built-in/src/app.js
@@ -1,0 +1,10 @@
+import tmpl from './views/app.hbs';
+
+const locals = {
+  title: 'Demo',
+  name: 'World',
+};
+
+document.getElementById('main').innerHTML = tmpl(locals);
+
+console.log('>> app');