+++ test/preprocessor.test.js
@@ -40,7 +40,7 @@ describe('handlebars', () => {
   test('access @root', () => compareFiles('_preprocessor/handlebars-access-root-variable'));
 
   // test helpers
-  test('build-in `block` helper', () => compareFiles('_preprocessor/handlebars-helper-block-buildIn'));
+  test('build-in `block` helper', () => compareFiles('_preprocessor/handlebars-helper-block-built-in'));
   test('override build-in helper', () => compareFiles('_preprocessor/handlebars-helper-block-user'));
   test('handlebars-layouts', () => compareFiles('_preprocessor/handlebars-helper-layouts'));
 
@@ -101,6 +101,7 @@ describe('usage template in js on client side', () => {
   test('hbs: compile to fn', () => compareFiles('_preprocessor/js-tmpl-hbs-compile'));
   test('hbs: compile undefined vars', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-strict-undefined-var'));
   test('hbs: compile, helpers', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-helpers'));
+  test('hbs: compile, built-in helpers', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-helpers-built-in'));
   test('hbs: compile, helpers, strict', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-helpers-strict'));
   test('hbs: compile, partials', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-partials'));
   test('hbs: compile, variables', () => compareFiles('_preprocessor/js-tmpl-hbs-compile-variables'));